Response::FindServer now returns a reference
As the first server of the corresponding host:port has to response if no
server is corresponding (because of server_names), it is certain to find
a Server. We can use references everywhere now !

Except in Response, where for the moment, it is possible to send a
reponse if no Request has been parsed. To be refactored.